refactor: Task, Active, Completed, Saving
Changed how we save tasks to database and ensured that the schemas are working as expected.

Moved task2 to the task file and just removed the old task code all together.

Resolved some issues with the ActiveTask showing weird numbers in progress

Added progress to completed tasks - seems to be broken still